third_party/blink/web_tests/external/wpt/periodic-background-sync/idlharness.https.any-expected.txt

This is a testharness.js-based test.
[FAIL] PeriodicSyncManager must be primary interface of registration.periodicSync
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: registration is not defined"
[FAIL] Stringification of registration.periodicSync
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: registration is not defined"
[FAIL] PeriodicSyncManager interface: registration.periodicSync must inherit property "register(DOMString, optional BackgroundSyncOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: registration is not defined"
[FAIL] PeriodicSyncManager interface: calling register(DOMString, optional BackgroundSyncOptions) on registration.periodicSync with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: registration is not defined"
[FAIL] PeriodicSyncManager interface: registration.periodicSync must inherit property "getTags()" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: registration is not defined"
[FAIL] PeriodicSyncManager interface: registration.periodicSync must inherit property "unregister(DOMString)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: registration is not defined"
[FAIL] PeriodicSyncManager interface: calling unregister(DOMString) on registration.periodicSync with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: registration is not defined"
[FAIL] PeriodicSyncEvent must be primary interface of new PeriodicSyncEvent("tag")
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: PeriodicSyncEvent is not defined"
[FAIL] Stringification of new PeriodicSyncEvent("tag")
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: PeriodicSyncEvent is not defined"
[FAIL] PeriodicSyncEvent interface: new PeriodicSyncEvent("tag") must not have property "undefined"
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: PeriodicSyncEvent is not defined"
[FAIL] PeriodicSyncEvent interface: new PeriodicSyncEvent("tag") must not have property "tag"
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: PeriodicSyncEvent is not defined"
[FAIL] ServiceWorkerRegistration interface: registration must inherit property "periodicSync" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: registration is not defined"
[FAIL] ServiceWorkerGlobalScope interface: onperiodicsync must not have property "onperiodicsync"
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: onperiodicsync is not defined"
Harness: the test ran to completion.
